Fixed going left on upgrade screen not working
Fixed going left on upgrade screen not working by changing postfix on GetUpgradableComponents to a transpiler. the m_CachedUpgradableComponents in the original method was only updated once per second and then for the rest of the time just returned without modification. The original postfix just constantly added duplicates of the hullplating and autoturrets every time leading to the cached components list becoming extremely bloated over time and reseting every second meaning that when you went left from 0 -> max of the list it sent you to index like 60 or 70 and then the on the second that the list reset the length of the list became 17 or so and then the engi UI would send you back to 0 because obviously if your over the max length of the list it should send you back to 0